requirements.txt
google-genai
numpy
faiss-cpu
flask 
flask-cors